Exporta informacões do usuário para componente ArtistInfo
